requirements.txt
Flask
python-dotenv
firebase-admin
Werkzeug
gunicorn